refactor: 용어 변경 -> articles
