+++ src/pages/Articles/hooks/useArticle.ts
@@ -1,5 +1,5 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
-import { notice } from 'api';
+import { articles } from 'api';
 
 const useArticle = (id: string | undefined) => {
   const { data: article } = useSuspenseQuery(
@@ -8,7 +8,7 @@ const useArticle = (id: string | undefined) => {
       queryFn: async ({ queryKey }) => {
         const queryFnParams = queryKey[1];
 
-        return notice.getArticle(queryFnParams);
+        return articles.getArticle(queryFnParams);
       },
     },
   );
