+++ src/pages/Articles/hooks/useDeleteLostItemArticles.ts
@@ -1,6 +1,6 @@
 import { isKoinError, sendClientError } from '@bcsdlab/koin';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { deleteLostItemArticle } from 'api/notice';
+import { deleteLostItemArticle } from 'api/articles';
 import useTokenState from 'utils/hooks/state/useTokenState';
 import showToast from 'utils/ts/showToast';
