+++ src/pages/Articles/utils/transform.ts
@@ -1,4 +1,4 @@
-import { LostItemArticlesRequestDTO, LostItemArticlesResponseDTO, SingleLostItemArticleResponseDTO } from 'api/notice/entity';
+import { LostItemArticlesRequestDTO, LostItemArticlesResponseDTO, SingleLostItemArticleResponseDTO } from 'api/articles/entity';
 import { LostItemArticlesForGet, LostItemArticlesForPost, SingleLostItemArticle } from 'pages/Articles/ts/types';
 
 export const transformLostItemArticles = (dto: LostItemArticlesResponseDTO)
